+++ ouds_core/lib/components/control/ouds_control_item.dart
@@ -1,3 +1,14 @@
+// Software Name: OUDS Flutter
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Software description: Flutter library of reusable graphical components
+//
+
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:ouds_core/components/control/internal/controller/ouds_interaction_state_controller.dart';
@@ -7,6 +18,12 @@ import 'package:ouds_core/components/control/internal/modifier/ouds_control_bord
 import 'package:ouds_core/components/control/internal/modifier/ouds_control_text_modifier.dart';
 import 'package:ouds_core/components/control/internal/ouds_control_state.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
+
+enum OudsControlItemType {
+  switchButton,
+  checkbox,
+  radio,
+}
 
 /// Refactor of controls for [Checkbox], [Switch], and [RadioButton].
 /// This implementation provides a customizable control item with properties such as text, icon, and interaction states.
@@ -22,6 +39,7 @@ class OudsControlItem extends StatefulWidget {
   final bool readOnly;
   final bool error;
   final String errorComponentName;
+  final OudsControlItemType componentType;
   final Widget Function() indicator;
   final String? additionalText;
 
@@ -31,6 +49,7 @@ class OudsControlItem extends StatefulWidget {
     super.key,
     required this.text,
     required this.errorComponentName,
+    required this.componentType,
     required this.indicator,
     this.helperText,
     this.icon,
@@ -136,7 +155,7 @@ class OudsControlItemState extends State<OudsControlItem> {
                       ),
                       child: IntrinsicHeight(
                         child: Row(
-                          crossAxisAlignment: CrossAxisAlignment.start,
+                          crossAxisAlignment: CrossAxisAlignment.center,
                           children: !widget.reversed ? _buildStandardLayout(controlItemState) : _buildInvertedLayout(controlItemState),
                         ),
                       ),
@@ -185,8 +204,12 @@ class OudsControlItemState extends State<OudsControlItem> {
             ),
             alignment: Alignment.center,
             child: SizedBox(
-              height: OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader,
-              width: OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader,
+              height: widget.componentType != OudsControlItemType.switchButton
+                  ? OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader
+                  : null,
+              width: widget.componentType != OudsControlItemType.switchButton
+                  ? OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader
+                  : null,
               child: widget.indicator(),
             ),
           ),
@@ -247,8 +270,12 @@ class OudsControlItemState extends State<OudsControlItem> {
             ),
             alignment: Alignment.center,
             child: SizedBox(
-              height: OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader,
-              width: OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader,
+              height: widget.componentType != OudsControlItemType.switchButton
+                  ? OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader
+                  : null,
+              width: widget.componentType != OudsControlItemType.switchButton
+                  ? OudsTheme.of(context).componentsTokens(context).controlItem.sizeLoader
+                  : null,
               child: widget.indicator(),
             ),
           ),
@@ -257,44 +284,46 @@ class OudsControlItemState extends State<OudsControlItem> {
 
   Widget _buildTextWithAdditionalAndHelper(OudsControlState controlItemState) {
     final controlItemTextModifier = OudsControlTextModifier(context);
+    final hasAdditionalText = widget.additionalText?.trim().isNotEmpty ?? false;
+    final hasHelperText = widget.helperText?.trim().isNotEmpty ?? false;
+
+    final List<Widget> columnChildren = [
+      Text(
+        widget.text,
+        style: OudsTheme.of(context).typographyTokens.typeLabelDefaultLarge(context).copyWith(
+              color: controlItemTextModifier.getTextColor(controlItemState, widget.error),
+            ),
+      ),
+    ];
+    if (hasAdditionalText) {
+      columnChildren.add(SizedBox(height: OudsTheme.of(context).componentsTokens(context).controlItem.spaceRowGap));
+      columnChildren.add(
+        Text(
+          widget.additionalText!,
+          style: OudsTheme.of(context).typographyTokens.typeLabelStrongMedium(context).copyWith(
+                color: controlItemTextModifier.getAdditionalTextColor(controlItemState),
+              ),
+        ),
+      );
+    }
+
+    if (hasHelperText) {
+      columnChildren.add(SizedBox(height: OudsTheme.of(context).componentsTokens(context).controlItem.spaceRowGap));
+      columnChildren.add(
+        Text(
+          widget.helperText!,
+          style: OudsTheme.of(context).typographyTokens.typeLabelDefaultMedium(context).copyWith(
+                color: controlItemTextModifier.getHelperTextColor(controlItemState),
+              ),
+        ),
+      );
+    }
+
     return Expanded(
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            widget.text,
-            style: TextStyle(
-              fontSize: OudsTheme.of(context).fontTokens.sizeLabelLarge,
-              letterSpacing: OudsTheme.of(context).fontTokens.letterSpacingLabelLarge,
-              fontWeight: OudsTheme.of(context).fontTokens.weightLabelDefault,
-              color: controlItemTextModifier.getTextColor(controlItemState, widget.error),
-            ),
-          ),
-          if (widget.additionalText != null) ...[
-            SizedBox(height: OudsTheme.of(context).componentsTokens(context).controlItem.spaceRowGap),
-            Text(
-              widget.additionalText!,
-              style: TextStyle(
-                fontSize: OudsTheme.of(context).fontTokens.sizeLabelMedium,
-                letterSpacing: OudsTheme.of(context).fontTokens.letterSpacingLabelMedium,
-                fontWeight: OudsTheme.of(context).fontTokens.weightStrong,
-                color: controlItemTextModifier.getAdditionalTextColor(controlItemState),
-              ),
-            ),
-          ],
-          if (widget.helperText != null) ...[
-            SizedBox(height: OudsTheme.of(context).componentsTokens(context).controlItem.spaceRowGap),
-            Text(
-              widget.helperText!,
-              style: TextStyle(
-                fontSize: OudsTheme.of(context).fontTokens.sizeLabelMedium,
-                letterSpacing: OudsTheme.of(context).fontTokens.letterSpacingLabelMedium,
-                fontWeight: OudsTheme.of(context).fontTokens.weightBodyDefault,
-                color: controlItemTextModifier.getHelperTextColor(controlItemState),
-              ),
-            ),
-          ],
-        ],
+        mainAxisSize: MainAxisSize.min, // prevents taking full height
+        children: columnChildren,
       ),
     );
   }